requirements.txt
opencv-python
mediapipe
numpy
streamlit
pandas
matplotlib
docker
fastapi
pymongo
PyYAML